iris: split launcher into package modules
