+++ hpc/iris/__init__.py
@@ -1,0 +1,5 @@
+"""Iris launcher support for OpenThoughts-Agent."""
+
+from hpc.iris.launcher import IrisLauncher
+
+__all__ = ["IrisLauncher"]
