+++ hpc/iris/settings.py
@@ -1,0 +1,5 @@
+"""Shared constants for OT-Agent Iris launchers."""
+
+DEFAULT_TASK_IMAGE = "ghcr.io/open-thoughts/openthoughts-agent:tpu"
+DEFAULT_CLUSTER_CONFIG = "lib/iris/config/marin.yaml"
+DEFAULT_PRIORITY = "interactive"
